Split admin_site from core to another app named admin_site
