+++ renthub/urls.py
@@ -14,10 +14,9 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from core.admin import admin_site
 from django.urls import path, include
 
 urlpatterns = [
     path('', include('core.urls')),
-    path('admin/', admin_site.urls),
+    path('', include('admin_site.urls'))
 ]
